Get more tests passing
[ci skip]
